Adding all project logic and tests

+++ Core/Model/SearchResult.cs
@@ -6,6 +6,6 @@ namespace SearchFight.Core.Model
 {
     public class SearchResult
     {
-        public int NumberOfResults { get; set; }
+        public long NumberOfResults { get; set; }
     }
 }
